Move sasl sub-package into internal/.
