+++ internal/sasl/sspi_windows.c
