+++ internal/sasl/sspi_windows.h
